requirements.txt
email-validator==1.2.1
fastapi==0.78.0
pandas==1.4.2
python-decouple==3.6
python-multipart==0.0.5
uvicorn==0.17.6
